Implement the buildNextMoveTree in Knight class
The Node class was also altered to have a children property which is
an array that holds other Nodes.

The Knight class was also given an array that holds all the next
possible moves in array form.

A Board class was created as well which is eventually become objects
that are passed into the buildNextMoveTree method to make sure
the a knight's next moves are within the boundaries of a user-defined
board size.